+++ tests/generic/models.py
@@ -154,6 +154,11 @@ class Person(models.Model):
         geom_collection = models.GeometryCollectionField()
 
 
+class ProxyToPersonModel(Person):
+    class Meta:
+        proxy = True
+
+
 class Dog(models.Model):
     class Meta:
         order_with_respect_to = "owner"
@@ -301,6 +306,18 @@ if BAKER_CONTENTTYPES:
         )
         object_id = models.PositiveIntegerField()
         content_object = GenericForeignKey("content_type", "object_id")
+
+        proxy_content_type = models.ForeignKey(
+            contenttypes.ContentType,
+            on_delete=models.CASCADE,
+            blank=True,
+            null=True,
+            limit_choices_to={"model__in": ["person", "dog"]},
+        )
+        proxy_object_id = models.PositiveIntegerField()
+        proxy_content_object = GenericForeignKey(
+            "proxy_content_type", "proxy_object_id", for_concrete_model=False
+        )
 
     class DummyGenericRelationModel(models.Model):
         relation = GenericRelation(DummyGenericForeignKeyModel)